+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.7.1
-* Removed after_help
-* Added `-f/--force` when getting a file
+* Removed after_help.
+* Added `-f/--force` when getting a file.
+* Replaced `-H/--head` with `-m/--meta` to retrieve metadata.
 
 ## 0.7.0 🚜
 * Added support for Additional Checksum Algorithms `x-amz-checksum-mode` option `--checksum`